requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Limiter==3.5.0
pytest==8.0.0
PyMySQL==1.1.0


